Splic logic into a pandas and xarray module
