+++ qcodes/dataset/exporters/__init__.py
@@ -1,0 +1,4 @@
+"""
+Functions for exporting data within a dataset or its cache to Pandas and Xarray formats.
+These functions are not considered part of the public api and may change and any time.
+"""
